Include indentation in line length calculation
This unfortunately makes the tradeoff that very deeply nested structures will end up "crunched up" near the edge of the column limit, but it is likely the desired behavior under most circumstances because it keeps things within a precise column limit.  It may be helpful to add a parameter to toggle this behavior in a future version.